Moved testing OpModes to proper directory, fixed some errors with parameters, first use of GamepadWrapper.
