+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/tests/LimelightTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.tests;
+package org.firstinspires.ftc.teamcode.opmodes.tests;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
@@ -37,7 +37,7 @@ public class LimelightTest extends LinearOpMode {
             telemetry.update();
             printStatus(limelight.getStatus());
 
-            driveBase.setVelocity(gamepad1.left_stick_y, -gamepad1.left_stick_x, -gamepad1.right_stick_x, 1);
+            driveBase.setVelocity(gamepad1.left_stick_y, -gamepad1.left_stick_x, -gamepad1.right_stick_x);
 
             LLResult latest = limelight.getLatestResult();
